Move modal cleanup fixtures to importable imbue.mng_modal.conftest
Move _reset_modal_app_registry and modal_session_cleanup from the
project-level conftest to the package-level conftest so consuming
packages (like mng_claude) get them via pytest_plugins import.
